Move Blob Storage service to services module
